+++ src/components/WorkOrders/CloseWorkOrderForm.js
@@ -14,8 +14,13 @@ import {
   OVERTIME_PAYMENT_TYPE,
   optionsForPaymentType,
 } from '../../utils/paymentTypes'
-import { CLOSURE_STATUS_OPTIONS } from '@/utils/statusCodes'
+import {
+  CLOSURE_STATUS_OPTIONS,
+  FOLLOW_ON_STATUS_OPTIONS,
+} from '@/utils/statusCodes'
 import { useState } from 'react'
+import FollowOnRequestTypeOfWorkForm from './FollowOnRequestTypeOfWorkForm'
+import FollowOnRequestMaterialsForm from './FollowOnRequestMaterialsForm'
 
 const CloseWorkOrderForm = ({
   reference,
@@ -28,7 +33,6 @@ const CloseWorkOrderForm = ({
   completionDate,
   startTime,
   startDate,
-  reason,
   dateRaised,
   selectedPercentagesToShowOnEdit,
   totalSMV,
@@ -43,6 +47,9 @@ const CloseWorkOrderForm = ({
     errors,
     trigger,
     getValues,
+    watch,
+    clearErrors,
+    setError,
   } = useForm({})
 
   const [startTimeIsRequired, setStartTimeIsRequired] = useState(false)
@@ -52,6 +59,12 @@ const CloseWorkOrderForm = ({
     setStartTimeIsRequired(e.target.value !== '')
   }
 
+  const showFollowOnRadioOptions = watch('reason') === 'Work Order Completed'
+  // const selectedFurtherWorkRequired =
+  //   watch('followOnStatus') === 'furtherWorkRequired'
+  const selectedFurtherWorkRequired =
+    watch('followOnStatus') === 'furtherWorkRequired'
+
   return (
     <div>
       <BackButton />
@@ -59,20 +72,58 @@ const CloseWorkOrderForm = ({
 
       <form role="form" onSubmit={handleSubmit(onSubmit)}>
         <Radios
+          labelSize="s"
           label="Select reason for closing"
           name="reason"
-          options={CLOSURE_STATUS_OPTIONS.map((r) => {
-            return {
-              text: r.text,
-              value: r.value,
-              defaultChecked: r.value === reason,
-            }
-          })}
+          options={CLOSURE_STATUS_OPTIONS.map((r) => ({
+            ...r,
+            children:
+              r.value === 'Work Order Completed' && showFollowOnRadioOptions ? (
+                <Radios
+                  name="followOnStatus"
+                  options={FOLLOW_ON_STATUS_OPTIONS}
+                  register={register({
+                    required: 'Please confirm if further work is required',
+                  })}
+                  error={errors && errors.followOnStatus}
+                />
+              ) : null,
+          }))}
           register={register({
             required: 'Please select a reason for closing the work order',
           })}
           error={errors && errors.reason}
         />
+
+        <div
+          style={{
+            display: selectedFurtherWorkRequired ? 'block' : 'none',
+          }}
+        >
+          <h1 className="lbh-heading-h2">Details of further work required</h1>
+
+          <FollowOnRequestTypeOfWorkForm
+            errors={errors}
+            register={register}
+            getValues={getValues}
+            setError={setError}
+            clearErrors={clearErrors}
+            watch={watch}
+          />
+
+          <FollowOnRequestMaterialsForm
+            register={register}
+            getValues={getValues}
+            errors={errors}
+          />
+
+          <TextArea
+            name="additionalNotes"
+            label="Additional notes"
+            register={register}
+            error={errors && errors.additionalNotes}
+          />
+        </div>
 
         {/* Start time cannot be changed once set by an operative */}
         {!existingStartTime && (
